updates on 20231129

+++ src/graph/dfs.hpp
@@ -1,7 +1,7 @@
 #include "../macros.hpp"
 
 // 木DP用定型部分
-auto dfs = [&](auto&& f, ll u, ll prv) -> void {
+auto dfs = [&](auto&& f, int u, int prv) -> void {
     for (auto v : nodes[u]) {
         if (v == prv) continue;
         f(f, v, u);
